requirements.txt
astroid==2.9.0
certifi==2022.6.15
click==8.1.3
dnslib==0.9.19
dnspython==2.2.1
Flask==2.0.2
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
PyJWT==2.4.0
pymongo==4.1.1
PythonDNS==0.1
Werkzeug==2.1.2
wrapt==1.13.3